Refactor CollectionPictureDisplay tests to BrowseView tests
Thereby simplify them and adjust associated mocks
